[CFDP-927] My dream teams feature (#135)
* Added algorithm for my dream team

* Just a thought for VolunteerGroup schema

* Updated interface naming to GroupItem to match the ContentItem naming schema

* Updated the VolunteerGroup resolver

* GroupFilterInput

* Updates Group Feature to exclude Dream Team groups

* Update Feature Flags

* Fixed Group Id typename resolver

* Create a GroupItem datasource for resolving VolunteerGroup requests
