+++ src/data/index.js
@@ -51,6 +51,7 @@ import * as Event from './events';
 import * as Feature from './features';
 import * as Flag from './flag';
 import * as Group from './groups';
+import * as GroupItem from './group-item'
 import * as LiveStream from './live-stream';
 import * as MatrixItem from './matrix-item';
 import * as Message from './message';
@@ -91,6 +92,7 @@ const data = {
   Flag,
   Followings,
   Group,
+  GroupItem,
   Interactions,
   LiveStream,
   MatrixItem,
